requested changes of PR
